Sugar 0.4.0: Server SSH key verification
Sugar now consults the server console to detect the SSH hostkey fingerprints, stores them in an EC2 tag for future reference, and adds the keys to the local system to bypass the yes/no prompt while improving security. It also auto-detects the username to use when connecting to the server.

This feature is still somewhat experimental.